refactored amazon price and load
