+++ openlibrary/catalog/add_book/__init__.py
@@ -36,7 +36,6 @@ from infogami import config
 
 from openlibrary.catalog.merge.merge_marc import build_marc
 from openlibrary.catalog.utils import mk_norm
-from openlibrary.core.vendors import get_amazon_metadata
 from openlibrary.core import lending
 from openlibrary.catalog.utils import flip_name
 from openlibrary import accounts
@@ -516,34 +515,6 @@ def update_ia_metadata_for_ol_edition(edition_id):
                 else:
                     data = item.metadata
     return data
-
-def create_edition_from_amazon_metadata(isbn):
-    """Fetches amazon metadata by isbn from affiliates API, attempts to
-    create OL edition from metadata, and returns the resulting edition key
-    `/key/OL..M` if successful or None otherwise
-    """
-    md = get_amazon_metadata(isbn)
-    if md:
-        reply = load_from_amazon_metadata(md)
-        if reply and reply.get('success'):
-            return reply['edition']['key']
-
-def load_from_amazon_metadata(rec):
-    """This is a bootstrapping helper method which enables us to take the
-    results of plugins.upstream.code.get_amazon_metadata and create an
-    OL book catalog record
-    """
-    
-    conforming_fields = [
-        'title', 'authors', 'publish_date', 'source_records',
-        'number_of_pages', 'publishers', 'cover', 'isbn_10',
-        'isbn_13']
-    conforming_rec = {}
-    for k in conforming_fields:
-        # if valid key and value not None
-        if rec.get(k) is not None:
-            conforming_rec[k] = rec[k]
-    return load(conforming_rec)
 
 def load(rec):
     """Given a record, tries to add/match that edition in the system.
